Fix EAD exporter
Replace / or space characters in unitid with _ for EAD filename
in zip.

Also move request handling into module that is included by
ead_export and can be included in other scripts as needed.
